requirements.txt
black==22.3.0
datapane==0.16.1
flake8==4.0.1
isort==5.12.0
mypy==1.1.1
pandas_datareader==0.10.0
pandasql==0.7.3
pushbullet.py==0.12.0
pylint==2.13.9
pytest==7.1.2
python-dotenv==0.20.0
psycopg2-binary==2.8.6
scipy==1.7.3
SQLAlchemy==1.4.36
yfinance==0.1.70
